style: Rename integration dir to integrations (plural)
