+++ internal/integrations/destination/example.go
